Replace long by Long.

+++ app/src/main/java/com/pokedeck/imie/pokedeck/entity/User.java
@@ -20,7 +20,7 @@ import java.util.ArrayList;
 
 public class User {
 
-    private long id;
+    private Long id;
 
     private String username;
 
@@ -39,7 +39,7 @@ public class User {
     public User() {
     }
 
-    public User(Integer id, String username, String email, String password, Boolean fighting, Pokedeck pokedeck, Pokemon currentPokemon, ArrayList<Pokemon> pokemonTeam) {
+    public User(Long id, String username, String email, String password, Boolean fighting, Pokedeck pokedeck, Pokemon currentPokemon, ArrayList<Pokemon> pokemonTeam) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -113,11 +113,11 @@ public class User {
         opponent.setFighting(true);
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
